create actions directory
