+++ app/api/article/route.ts
@@ -9,99 +9,6 @@ const createArticleSchema = z.object({
   tags: z.array(z.string()),
 });
 
-export async function fetchAllArticles() {
-  try {
-    const articles = await prisma.article.findMany({
-      orderBy: {
-        createdAt: 'desc',
-      },
-      include: {
-        author: {
-          select: {
-            name: true,
-          },
-        },
-        _count: {
-          select: {
-            claps: true,
-          },
-        },
-      },
-    });
-    return articles;
-  } catch (error) {
-    throw new Error(String(error));
-  }
-}
-
-export async function fetchArticleById(id: number) {
-  try {
-    const article = await prisma.article.findUnique({
-      where: {
-        id,
-      },
-      include: {
-        author: {
-          select: {
-            name: true,
-          },
-        },
-        comments: {
-          orderBy: {
-            createdAt: 'desc',
-          },
-          select: {
-            body: true,
-            createdAt: true,
-            authorId: true,
-            author: {
-              select: {
-                name: true,
-              },
-            },
-          },
-        },
-        _count: {
-          select: {
-            claps: true,
-          },
-        },
-      },
-    });
-    return article;
-  } catch (error) {
-    throw new Error(String(error));
-  }
-}
-
-export async function fetchArticlesByAuthorId(authorId: number) {
-  try {
-    const articles = await prisma.article.findMany({
-      where: {
-        authorId,
-      },
-      orderBy: {
-        createdAt: 'desc',
-      },
-      include: {
-        author: {
-          select: {
-            name: true,
-          },
-        },
-        _count: {
-          select: {
-            claps: true,
-          },
-        },
-      },
-    });
-    return articles;
-  } catch (error) {
-    throw new Error(String(error));
-  }
-}
-
 // POST - Create a new article
 // Path: /api/article
 export async function POST(request: NextRequest) {
